requirements.txt
streamlit
supabase
anthropic
httpx
python-dotenv
requests
feedparser
